# internal/adapter/http/handler/client_project/common_dto.go
package client_project

import "time"

// ProjectResponse is the common response DTO for project
type ProjectResponse struct {
	ProjectID   string    `json:"project_id"`
	Name        string    `json:"name"`
	Description string    `json:"description"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
	NamespaceID string    `json:"namespace_id"`
}